Move VisionBlockingAccumulator into its own file
`FogUtil` is getting cluttered enough, so `VisionBlockingAccumulator` can be its own thing now.
