+++ src/main/java/net/rptools/maptool/client/ui/zone/FogUtil.java
@@ -30,11 +30,9 @@ import javax.annotation.Nullable;
 import net.rptools.lib.CodeTimer;
 import net.rptools.maptool.client.AppUtil;
 import net.rptools.maptool.client.MapTool;
-import net.rptools.maptool.client.ui.zone.vbl.AreaContainer;
-import net.rptools.maptool.client.ui.zone.vbl.AreaIsland;
-import net.rptools.maptool.client.ui.zone.vbl.AreaOcean;
 import net.rptools.maptool.client.ui.zone.vbl.AreaTree;
 import net.rptools.maptool.client.ui.zone.vbl.VisibleAreaSegment;
+import net.rptools.maptool.client.ui.zone.vbl.VisionBlockingAccumulator;
 import net.rptools.maptool.model.AbstractPoint;
 import net.rptools.maptool.model.CellPoint;
 import net.rptools.maptool.model.ExposedAreaMetaData;
@@ -76,7 +74,7 @@ public class FogUtil {
     vision.transform(AffineTransform.getTranslateInstance(x, y));
 
     Point origin = new Point(x, y);
-    var accumulator = new VisionBlockingAccumulator(origin);
+    var accumulator = new VisionBlockingAccumulator(geometryFactory, origin);
     if (!accumulator.addWallBlocking(topology)
         || !accumulator.addHillBlocking(hillVbl)
         || !accumulator.addPitBlocking(pitVbl)) {
@@ -441,181 +439,5 @@ public class FogUtil {
     y = bounds.y + bounds.height / 2;
 
     return new Point(x, y);
-  }
-
-  private static final class VisionBlockingAccumulator {
-    private final Point origin;
-    private final List<VisibleAreaSegment> visionBlockingSegments;
-
-    public VisionBlockingAccumulator(Point origin) {
-      this.origin = origin;
-      this.visionBlockingSegments = new ArrayList<>();
-    }
-
-    public List<VisibleAreaSegment> getVisionBlockingSegments() {
-      return visionBlockingSegments;
-    }
-
-    private void addVisionBlockingSegments(AreaContainer areaContainer, boolean frontSide) {
-      var segments =
-          areaContainer.getVisionBlockingBoundarySegements(geometryFactory, origin, frontSide);
-      visionBlockingSegments.addAll(segments);
-    }
-
-    private void addIslandForHillBlocking(
-        AreaIsland blockingIsland, @Nullable AreaOcean excludeChildOcean) {
-      // The back side of the island blocks.
-      addVisionBlockingSegments(blockingIsland, false);
-      // The front side of each contained ocean also acts as a back side boundary of the island.
-      for (var blockingOcean : blockingIsland.getOceans()) {
-        if (blockingOcean == excludeChildOcean) {
-          continue;
-        }
-
-        addVisionBlockingSegments(blockingOcean, true);
-      }
-    }
-
-    /**
-     * Finds all wall topology segments that can take part in blocking vision.
-     *
-     * @param topology The topology to treat as Wall VBL.
-     * @return false if the vision has been completely blocked by topology, or true if vision may be
-     *     blocked by particular segments.
-     */
-    public boolean addWallBlocking(AreaTree topology) {
-      final AreaContainer container = topology.getContainerAt(origin);
-      if (container == null) {
-        // Should never happen since the global ocean should catch everything.
-        return false;
-      }
-
-      if (container instanceof AreaIsland) {
-        // Since we're contained in a wall island, there can be no vision through it.
-        return false;
-      } else if (container instanceof AreaOcean ocean) {
-        final var parentIsland = ocean.getParentIsland();
-        if (parentIsland != null) {
-          // The near edge of the island blocks vision, which is the same as the boundary of this
-          // ocean. Since we're inside the ocean, the facing edges of the parent island are like the
-          // back side.
-          addVisionBlockingSegments(ocean, false);
-        }
-
-        // Check each contained island.
-        for (var containedIsland : ocean.getIslands()) {
-          // The front side of wall VBL blocks vision.
-          addVisionBlockingSegments(containedIsland, true);
-        }
-      }
-
-      return true;
-    }
-
-    /**
-     * Finds all hill topology segments that can take part in blocking vision.
-     *
-     * @param topology The topology to treat as Hill VBL.
-     * @return false if the vision has been completely blocked by topology, or true if vision can be
-     *     blocked by particular segments.
-     */
-    public boolean addHillBlocking(AreaTree topology) {
-      final AreaContainer container = topology.getContainerAt(origin);
-      if (container == null) {
-        // Should never happen since the global ocean should catch everything.
-        return false;
-      }
-
-      /*
-       * There are two cases for Hill VBL:
-       * 1. A token inside hill VBL can see into adjacent oceans, and therefore into other areas of
-       *    Hill VBL in those oceans.
-       * 2. A token outside hill VBL can see into hill VBL, but not into any oceans adjacent to it.
-       */
-
-      if (container instanceof final AreaIsland island) {
-        /*
-         * Since we're in an island, vision is blocked by:
-         * 1. The back side of the parent ocean's parent island.
-         * 2. The back side of any sibling islands (other children of the parent ocean).
-         * 3. The back side of any child ocean's islands.
-         * 4. For each island in the above, any child ocean provided it's not the parent of the
-         *    current island.
-         */
-        final var parentOcean = island.getParentOcean();
-
-        final var grandparentIsland = parentOcean.getParentIsland();
-        if (grandparentIsland != null) {
-          addIslandForHillBlocking(grandparentIsland, parentOcean);
-        }
-
-        for (final var siblingIsland : parentOcean.getIslands()) {
-          if (siblingIsland == island) {
-            // We don't want to block vision for the hill we're currently in.
-            // TODO Ideally we could block the second occurence of the current island, but we need
-            //  a way to do that reliably.
-            continue;
-          }
-
-          addIslandForHillBlocking(siblingIsland, null);
-        }
-
-        for (final var childOcean : island.getOceans()) {
-          for (final var grandchildIsland : childOcean.getIslands()) {
-            addIslandForHillBlocking(grandchildIsland, null);
-          }
-        }
-      } else if (container instanceof final AreaOcean ocean) {
-        /*
-         * Since we're in an ocean, vision is blocked by:
-         * 1. The back side of the parent island.
-         * 2. The back side of any child island
-         * 3. For each island in the above, any child ocean provided it's not the current ocean.
-         */
-        final var parentIsland = ocean.getParentIsland();
-        if (parentIsland != null) {
-          addIslandForHillBlocking(parentIsland, null);
-        }
-        // Check each contained island.
-        for (var containedIsland : ocean.getIslands()) {
-          addIslandForHillBlocking(containedIsland, null);
-        }
-      }
-
-      return true;
-    }
-
-    /**
-     * Finds all pit topology segments that can take part in blocking vision.
-     *
-     * @param topology The topology to treat as Pit VBL.
-     * @return false if the vision has been completely blocked by topology, or true if vision can be
-     *     blocked by particular segments.
-     */
-    public boolean addPitBlocking(AreaTree topology) {
-      final AreaContainer container = topology.getContainerAt(origin);
-      if (container == null) {
-        // Should never happen since the global ocean should catch everything.
-        return false;
-      }
-
-      /*
-       * There are two cases for Pit VBL:
-       * 1. A token inside Pit VBL can see only see within the current island, not into any adjacent
-       *    oceans.
-       * 2. A token outside Pit VBL is unobstructed by the Pit VBL (nothing special to do).
-       */
-
-      if (container instanceof final AreaIsland island) {
-        /*
-         * Since we're in an island, vision is blocked by:
-         * 1. The back side of the island.
-         * 2. The front side of any child ocean.
-         */
-        addIslandForHillBlocking(island, null);
-      }
-
-      return true;
-    }
   }
 }
